Added function to Travel window
The button Details is now in function that opens a Travel Details window.

the window also got some design done with some labels, listview and a button.

TODO: Add some travels for user to add to a list with some information.
Display those travels in details window.
and make remove button function work.