Big changes to program
Alot of code was change due to me(programmer) was getting stucked on the way as i was coding.

because I couldnt save anthing in user.travels.

after changing the code it would not work, however after checking all windows. found the fault that travelmanager was not set to new();
in mainwindow. which meant travelmanager was null all the time.

after 7-9 hours coding and trying to find all the bugs i can finally say good job to myself and have a good night rest!

TODO:
as i changed user = SignedInUser

need to change method variables in userdetails.

make Admin window

and to make details window show information about the tag from travelwindow